+++ TracRemote/tests/t/attach.html
@@ -71,7 +71,7 @@
               <dt>
     <a href="/attachment/wiki/WikiStart/sdss4collaborationmeeting_liuchao_20140727.key.zip" title="View attachment">sdss4collaborationmeeting_liuchao_20140727.key.zip</a><a href="/raw-attachment/wiki/WikiStart/sdss4collaborationmeeting_liuchao_20140727.key.zip" class="trac-rawlink" title="Download">​</a>
        (<span title="3097741 bytes">3.0 MB</span>) -
-      added by <em>liuchao</em> <a class="timeline" href="/timeline?from=2014-07-27T11%3A21%3A37-06%3A00&amp;precision=second" title="See timeline at 07/27/14 11:21:37">2 years ago</a>.
+      added by <span class="trac-author">liuchao</span> <a class="timeline" href="/timeline?from=2014-07-27T11%3A21%3A37-06%3A00&amp;precision=second" title="See timeline at 07/27/14 11:21:37">2 years ago</a>.
   </dt>
               <dd>
                 presentation file in SDSS IV collaboration meeting
@@ -79,17 +79,17 @@
               <dt>
     <a href="/attachment/wiki/WikiStart/jborissova_APOGEEII_28072017_final.zip" title="View attachment">jborissova_APOGEEII_28072017_final.zip</a><a href="/raw-attachment/wiki/WikiStart/jborissova_APOGEEII_28072017_final.zip" class="trac-rawlink" title="Download">​</a>
        (<span title="1861986 bytes">1.8 MB</span>) -
-      added by <em>jura</em> <a class="timeline" href="/timeline?from=2014-07-27T14%3A39%3A20-06%3A00&amp;precision=second" title="See timeline at 07/27/14 14:39:20">2 years ago</a>.
+      added by <span class="trac-author">jura</span> <a class="timeline" href="/timeline?from=2014-07-27T14%3A39%3A20-06%3A00&amp;precision=second" title="See timeline at 07/27/14 14:39:20">2 years ago</a>.
   </dt>
               <dt>
     <a href="/attachment/wiki/WikiStart/smao-apogee-2014.pdf" title="View attachment">smao-apogee-2014.pdf</a><a href="/raw-attachment/wiki/WikiStart/smao-apogee-2014.pdf" class="trac-rawlink" title="Download">​</a>
        (<span title="10799020 bytes">10.3 MB</span>) -
-      added by <em>shudemao</em> <a class="timeline" href="/timeline?from=2014-07-27T18%3A24%3A24-06%3A00&amp;precision=second" title="See timeline at 07/27/14 18:24:24">2 years ago</a>.
+      added by <span class="trac-author">shudemao</span> <a class="timeline" href="/timeline?from=2014-07-27T18%3A24%3A24-06%3A00&amp;precision=second" title="See timeline at 07/27/14 18:24:24">2 years ago</a>.
   </dt>
               <dt>
     <a href="/attachment/wiki/WikiStart/carigi.apogge2.lr.utah.pdf" title="View attachment">carigi.apogge2.lr.utah.pdf</a><a href="/raw-attachment/wiki/WikiStart/carigi.apogge2.lr.utah.pdf" class="trac-rawlink" title="Download">​</a>
        (<span title="4246601 bytes">4.0 MB</span>) -
-      added by <em>carigi</em> <a class="timeline" href="/timeline?from=2014-07-29T06%3A11%3A13-06%3A00&amp;precision=second" title="See timeline at 07/29/14 06:11:13">2 years ago</a>.
+      added by <span class="trac-author">carigi</span> <a class="timeline" href="/timeline?from=2014-07-29T06%3A11%3A13-06%3A00&amp;precision=second" title="See timeline at 07/29/14 06:11:13">2 years ago</a>.
   </dt>
               <dd>
                 carigi talk
@@ -97,7 +97,7 @@
               <dt>
     <a href="/attachment/wiki/WikiStart/carigi.apogge2.lr.utah.2.pdf" title="View attachment">carigi.apogge2.lr.utah.2.pdf</a><a href="/raw-attachment/wiki/WikiStart/carigi.apogge2.lr.utah.2.pdf" class="trac-rawlink" title="Download">​</a>
        (<span title="4246601 bytes">4.0 MB</span>) -
-      added by <em>carigi</em> <a class="timeline" href="/timeline?from=2014-07-29T06%3A14%3A44-06%3A00&amp;precision=second" title="See timeline at 07/29/14 06:14:44">2 years ago</a>.
+      added by <span class="trac-author">carigi</span> <a class="timeline" href="/timeline?from=2014-07-29T06%3A14%3A44-06%3A00&amp;precision=second" title="See timeline at 07/29/14 06:14:44">2 years ago</a>.
   </dt>
               <dd>
                 carigi talk
@@ -105,12 +105,12 @@
               <dt>
     <a href="/attachment/wiki/WikiStart/rcarrera_parkcity2014.pdf" title="View attachment">rcarrera_parkcity2014.pdf</a><a href="/raw-attachment/wiki/WikiStart/rcarrera_parkcity2014.pdf" class="trac-rawlink" title="Download">​</a>
        (<span title="884365 bytes">863.6 KB</span>) -
-      added by <em>rcarrera</em> <a class="timeline" href="/timeline?from=2014-07-29T09%3A33%3A23-06%3A00&amp;precision=second" title="See timeline at 07/29/14 09:33:23">2 years ago</a>.
+      added by <span class="trac-author">rcarrera</span> <a class="timeline" href="/timeline?from=2014-07-29T09%3A33%3A23-06%3A00&amp;precision=second" title="See timeline at 07/29/14 09:33:23">2 years ago</a>.
   </dt>
               <dt>
     <a href="/attachment/wiki/WikiStart/Thakar_CAS_DataReview.pdf" title="View attachment">Thakar_CAS_DataReview.pdf</a><a href="/raw-attachment/wiki/WikiStart/Thakar_CAS_DataReview.pdf" class="trac-rawlink" title="Download">​</a>
        (<span title="1942262 bytes">1.9 MB</span>) -
-      added by <em>thakar</em> <a class="timeline" href="/timeline?from=2014-10-13T12%3A41%3A57-06%3A00&amp;precision=second" title="See timeline at 10/13/14 12:41:57">22 months ago</a>.
+      added by <span class="trac-author">thakar</span> <a class="timeline" href="/timeline?from=2014-10-13T12%3A41%3A57-06%3A00&amp;precision=second" title="See timeline at 10/13/14 12:41:57">22 months ago</a>.
   </dt>
               <dd>
                 CAS presentation, Wednesday PM
